# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "FlaskExploreV1",
    "contact": {}
  },
  "host": "",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/login": {
      "post": {
        "description": "Login auth",
        "summary": "/login",
        "tags": [
          "auth"
        ],
        "operationId": "http://127.0.0.1:5000/login",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Accept",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "",
            "value": "application/json"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1127.0.0.1%3A5000~1loginrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/logout": {
      "delete": {
        "description": "logout",
        "summary": "/logout",
        "tags": [
          "auth"
        ],
        "operationId": "http://127.0.0.1:5000/logout",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Accept",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "",
            "value": "application/json"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
...
  "security": [],
...
}